refactor(job): 替换为MyBatis-Plus

+++ rabbit-job-common/src/main/java/com/snail/job/common/enums/TriggerStatus.java
@@ -10,19 +10,19 @@ public enum TriggerStatus {
     /**
      * 已停止
      */
-    STOPPED("stopped", (byte) 0, "已停止"),
+    STOPPED("stopped", 0, "已停止"),
 
     /**
      * 运行中
      */
-    RUNNING("running", (byte) 1, "运行中"),
+    RUNNING("running", 1, "运行中"),
     ;
 
     private final String name;
-    private final Byte value;
+    private final Integer value;
     private final String desc;
 
-    TriggerStatus(String name, Byte value, String desc) {
+    TriggerStatus(String name, Integer value, String desc) {
         this.name = name;
         this.value = value;
         this.desc = desc;
@@ -32,7 +32,7 @@ public enum TriggerStatus {
         return name;
     }
 
-    public Byte getValue() {
+    public Integer getValue() {
         return value;
     }
